// social-media/src/enums/social-media.enums.ts
export enum SocialMediaButtonsEnum {
    facebook = 'facebook',
    instagram = 'instagram',
    twitter = 'twitter',
    whatsapp = 'whatsapp',
    telegram = 'telegram',
    pintrest = 'pintrest',
    linkedin = 'linkedin',
    youtube = 'youtube',
    messenger = 'messenger',
    email = 'email',
};